Fix intersection computation
Previous implementation did not detect correctly intersecting elements
when they are a different indexes.

Changed the spec for the function to include duplicated elements as it
seems to be more natural, and a deduplicated iterator can be obtained using
`dedup` adaptor from `itertools` where as obtaining the duplicated
iterator from the deduplicated one is not easy.